backend/requirements.txt
# alembic==1.16.4
# blinker==1.9.0
# certifi==2025.8.3
# cffi==1.17.1
charset-normalizer==3.4.3
# click==8.2.1
cryptography==46.0.3
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.1
Flask-DebugToolbar
Flask-Login
Flask-Mail
Flask-Migrate
Flask-SQLAlchemy
Flask-WTF
# greenlet==3.2.4
# idna==3.10
# itsdangerous==2.2.0
jaconv
Jinja2
jpholiday
# Mako==1.3.10
# MarkupSafe==3.0.2
numpy==2.3.2
pandas==2.3.1
# pip==25.2
# pycparser==2.22
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
# pytz==2025.2
requests==2.32.4
# six==1.17.0
SQLAlchemy==2.0.43
# typing_extensions==4.14.1
# tzdata==2025.2
# urllib3==2.5.0
Werkzeug==3.1.3
WTForms==3.2.1